All CRUD done

+++ database/migrations/2025_11_03_230920_create_transaksis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('transaksis', function (Blueprint $table) {
             $table->id();
             $table->string('keterangan');
-            $table->decimal('nominal', 15, 2);
+            $table->integer('nominal');
             $table->enum('status', ['Masuk', 'Keluar', 'Withdraw']);
             $table->foreignId('kategori_id')->constrained('kategoris')->onDelete('cascade');
             $table->foreignId('dompet_id')->constrained('dompets')->onDelete('cascade');
